Tests/LibWeb/Text/expected/URL/search-params-with-invalid-utf8.txt

key equals   = true
value equals = true
